Adds ECMAScript modules
